chore: separate actions and types for axelar txs
