+++ packages/transaction-recovery/src/manual-relay-to-dest-chain/lib/AxelarRouteMessageTx/types.ts
@@ -1,0 +1,13 @@
+import {
+  AxelarRecoveryApiClient,
+  type SearchGMPResponseData,
+} from "@axelarjs/api";
+
+export type SendAxelarRouteMessageTxParams = {
+  messageId?: string | undefined;
+  searchGMPData: SearchGMPResponseData;
+};
+
+export type SendAxelarRouteMessageTxDependencies = {
+  axelarRecoveryApiClient: AxelarRecoveryApiClient;
+};
