+++ packages/transaction-recovery/src/manual-relay-to-dest-chain/lib/AxelarSignTx/index.ts
@@ -1,69 +1,3 @@
-import type {
-  AxelarRecoveryApiClient,
-  AxelarscanClient,
-  SearchGMPResponseData,
-} from "@axelarjs/api";
-
-import type { RecoveryTxResponse, ChainConfig } from "../../types";
-import { SignCommandsError, SignCommandsSkipReason } from "./constants";
-
-export type SendAxelarSignTxParams = {
-  searchGMPData: SearchGMPResponseData;
-  srcChainConfig: ChainConfig;
-};
-
-export type SendAxelarSignTxDependencies = {
-  axelarscanClient: AxelarscanClient;
-  axelarRecoveryApiClient: AxelarRecoveryApiClient;
-};
-
-export async function sendAxelarSignTx(
-  params: SendAxelarSignTxParams,
-  deps: SendAxelarSignTxDependencies
-): Promise<RecoveryTxResponse> {
-  const commandId = params.searchGMPData.command_id;
-  const sourceTransactionHash = params.searchGMPData.call.transactionHash;
-  const chainId = params.srcChainConfig.id;
-
-  const batchedCommands = await deps.axelarscanClient
-    .searchBatchedCommands({
-      commandId,
-      sourceTransactionHash,
-    })
-    .catch(() => ({ data: [] }));
-
-  // If there are batched commands, skip signing
-  if (batchedCommands?.data.length > 0) {
-    return {
-      skip: true,
-      type: "axelar.sign_commands",
-      skipReason: SignCommandsSkipReason.ALREADY_EXECUTED,
-    };
-  }
-
-  try {
-    const tx = await deps.axelarRecoveryApiClient.signCommands(chainId, "evm");
-
-    if (tx.code !== 0) {
-      return {
-        skip: true,
-        type: "axelar.sign_commands",
-        error: SignCommandsError.SIGN_COMMANDS_FAILED.message,
-      };
-    }
-
-    return {
-      skip: false,
-      type: "axelar.sign_commands",
-      tx: {
-        transactionHash: tx.transactionHash,
-      },
-    };
-  } catch (e) {
-    return {
-      skip: true,
-      type: "axelar.sign_commands",
-      error: SignCommandsError.SIGN_COMMANDS_FAILED.message,
-    };
-  }
-}
+export * from "./action";
+export * from "./types";
+export * from "./constants";
