+++ packages/transaction-recovery/src/manual-relay-to-dest-chain/lib/AxelarSignTx/types.ts
@@ -1,0 +1,17 @@
+import type {
+  AxelarRecoveryApiClient,
+  AxelarscanClient,
+  SearchGMPResponseData,
+} from "@axelarjs/api";
+
+import type { ChainConfig } from "../../types";
+
+export type SendAxelarSignTxParams = {
+  searchGMPData: SearchGMPResponseData;
+  srcChainConfig: ChainConfig;
+};
+
+export type SendAxelarSignTxDependencies = {
+  axelarscanClient: AxelarscanClient;
+  axelarRecoveryApiClient: AxelarRecoveryApiClient;
+};
